Fix for side-menu and scroll message when footer is displayed
